@@ -7,16 +7,12 @@ import (
 )
 
 func TestNewMsgCompressor(t *testing.T) {
-
 	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", 35.5, nil, nil, nil)
 
 	comp := NewMsgCompressor("", "")
 	decomp := NewMsgCompressor("", "")
 
 	for i := 0; i < 5; i++ {
-		bmsg, _ := msg.SerializeToJson()
-
-		t.Log("Uncompressed message size = ", len(bmsg))
 		compMsg, err := comp.CompressFimpMsg(msg)
 		if err != nil {
 			t.Fatal("Compressor error :", err)
@@ -27,9 +23,7 @@ func TestNewMsgCompressor(t *testing.T) {
 			t.Fatal("Compressor error 2 :", err)
 		}
 		if fimpMsg.Service != "temp_sensor" {
-			t.Fatal("Incorrect service name ")
-		} else {
-			t.Log("All good")
+			t.Fatal("Incorrect service name ", fimpMsg.Service)
 		}
 	}
 }
@@ -43,8 +37,7 @@ func TestNewMsgCompressor2(t *testing.T) {
 	comp := NewMsgCompressor("", "")
 	decomp := NewMsgCompressor("", "")
 
-	for i := 0; i < 10; i++ {
-
+	for range 10 {
 		msg, err := NewMessageFromBytes(bmsg0)
 		if err != nil {
 			t.Fatal("Deserialization error")
@@ -53,32 +46,33 @@ func TestNewMsgCompressor2(t *testing.T) {
 		if msg.ValueType == VTypeObject {
 			err := msg.GetObjectValue(&msg.Value)
 			if err != nil {
-				t.Fatal("<ses> Compression fimp error:", err.Error())
+				t.Fatal("Compression fimp error:", err.Error())
 			}
 		}
 		bmsg, _ := msg.SerializeToJson()
-		t.Log("Uncompressed message size = ", len(bmsg))
+
 		compMsg, err := comp.CompressBinMsg(bmsg)
 		if err != nil {
 			t.Fatal("Compressor error :", err)
 		}
-		t.Log("Compressed message size 1 = ", len(compMsg))
+
 		fimpMsgBin, err := decomp.DecompressBinMsg(compMsg)
 		if err != nil {
 			t.Fatal("Compressor error 1:", err)
 		}
-		//t.Log("Compressed message size 2 = ",len(fimpMsgBin))
+
 		fimpMsg, err := NewMessageFromBytes(fimpMsgBin)
 		//fimpMsg , err := decomp.DecompressFimpMsg(compMsg)
 		if err != nil {
 			t.Fatal("Compressor error 2:", err)
 		}
 		v1 := fimptype.ThingInclusionReport{}
-		fimpMsg.GetObjectValue(&v1)
+		if err := fimpMsg.GetObjectValue(&v1); err != nil {
+			t.Fatal("Getting object value error:", err)
+		}
+
 		if fimpMsg.Service != "test" || v1.HwVersion != "hw_version" {
-			t.Fatal("Incorrect service name ")
-		} else {
-			t.Log("All good")
+			t.Fatalf("Incorrect service name %s or hw version %s", fimpMsg.Service, v1.HwVersion)
 		}
 	}
 }
