@@ -5,20 +5,21 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"io"
+	"net/http"
+	"time"
+
 	"github.com/futurehomeno/fimpgo"
 	"github.com/futurehomeno/fimpgo/utils"
 	log "github.com/sirupsen/logrus"
-	"io/ioutil"
-	"net/http"
-	"time"
 )
 
 type OAuth2TokenResponse struct {
-	AccessToken  string      `json:"access_token"`
-	TokenType    string      `json:"token_type"`
-	ExpiresIn    int64       `json:"expires_in"`
-	RefreshToken string      `json:"refresh_token"`
-	Scope        interface{} `json:"scope"`
+	AccessToken  string `json:"access_token"`
+	TokenType    string `json:"token_type"`
+	ExpiresIn    int64  `json:"expires_in"`
+	RefreshToken string `json:"refresh_token"`
+	Scope        any    `json:"scope"`
 }
 
 type OAuth2RefreshProxyRequest struct {
@@ -73,9 +74,9 @@ func (oac *FhOAuth2Client) SetRefreshTokenApiUrl(refreshTokenApiUrl string) {
 	oac.refreshTokenApiUrl = refreshTokenApiUrl
 }
 
-//NewFhOAuth2Client implements OAuth client which communicates to 3rd party API over FH Auth proxy.
+// NewFhOAuth2Client implements OAuth client which communicates to 3rd party API over FH Auth proxy.
 func NewFhOAuth2Client(partnerName string, appName string, env string) *FhOAuth2Client {
-	client := &FhOAuth2Client{partnerName: partnerName, mqttServerURI: "tcp://localhost:1883", mqttClientID: "auth_client_" + appName}
+	client := &FhOAuth2Client{partnerName: partnerName, mqttServerURI: "tcp://127.0.0.1:1883", mqttClientID: "auth_client_" + appName}
 	if env == utils.EnvBeta {
 		client.refreshTokenApiUrl = "https://partners-beta.futurehome.io/api/control/edge/proxy/refresh"
 		client.authCodeApiUrl = "https://partners-beta.futurehome.io/api/control/edge/proxy/auth-code"
@@ -146,17 +147,20 @@ func (oac *FhOAuth2Client) LoadHubTokenFromCB() error {
 		}
 	}
 	responseTopic := fmt.Sprintf("pt:j1/mt:rsp/rt:app/rn:%s/ad:1", oac.appName)
-	oac.syncClient.AddSubscription(responseTopic)
+	if err := oac.syncClient.AddSubscription(responseTopic); err != nil {
+		return err
+	}
+
 	reqMsg := fimpgo.NewStringMessage("cmd.clbridge.get_auth_token", "clbridge", "", nil, nil, nil)
 	reqMsg.ResponseToTopic = responseTopic
 	var err error
 	var response *fimpgo.FimpMessage
-	for i := 0; i < oac.cbRetry; i++ {
+	for range oac.cbRetry {
 		response, err = oac.syncClient.SendFimp("pt:j1/mt:cmd/rt:app/rn:clbridge/ad:1", reqMsg, 5)
 		if err == nil {
 			break
 		}
-		log.Error("CB is not responding.Retrying")
+		log.Error("[edgeapp] CB is not responding.Retrying")
 		time.Sleep(time.Second * oac.cbRetryDelay)
 	}
 
@@ -184,9 +188,8 @@ func (oac *FhOAuth2Client) ExchangeRefreshToken(refreshToken string) (*OAuth2Tok
 	return oac.postMsg(req, oac.refreshTokenApiUrl)
 }
 
-func (oac *FhOAuth2Client) postMsg(req interface{}, url string) (*OAuth2TokenResponse, error) {
+func (oac *FhOAuth2Client) postMsg(req any, url string) (*OAuth2TokenResponse, error) {
 	if oac.hubToken == "" {
-		log.Info("Empty token.Re-requesting new token")
 		err := oac.LoadHubTokenFromCB()
 		if err != nil {
 			return nil, errors.New("empty hub token.operation aborted")
@@ -200,14 +203,14 @@ func (oac *FhOAuth2Client) postMsg(req interface{}, url string) (*OAuth2TokenRes
 	r, _ := http.NewRequest("POST", url, bytes.NewBuffer(reqB))
 	r.Header.Add("Content-Type", "application/json")
 	r.Header.Add("Authorization", "Bearer "+oac.hubToken)
-	//log.Info("Sending using token :",oac.hubToken
+
 	var resp *http.Response
-	for i := 0; i < oac.refreshRetry; i++ {
+	for range oac.refreshRetry {
 		resp, err = client.Do(r)
 		if err == nil && resp.StatusCode < 400 {
 			break
 		}
-		log.Error("Error response from auth endpoint.Retrying...")
+		log.Error("[fimpgo] Error response from auth endpoint.Retrying...")
 		time.Sleep(time.Second * oac.retryDelay)
 	}
 	if err != nil {
@@ -217,7 +220,7 @@ func (oac *FhOAuth2Client) postMsg(req interface{}, url string) (*OAuth2TokenRes
 		return nil, fmt.Errorf("error %s response from server", resp.Status)
 	}
 
-	bData, err := ioutil.ReadAll(resp.Body)
+	bData, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return nil, err
 	}
