@@ -3,28 +3,28 @@ package edgeapp
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/futurehomeno/fimpgo/utils"
-	log "github.com/sirupsen/logrus"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"time"
+
+	"github.com/futurehomeno/fimpgo/utils"
+	log "github.com/sirupsen/logrus"
 )
 
 type Configs struct {
 	path               string
-	WorkDir            string      `json:"-"`
-	InstanceAddress    string      `json:"instance_address"`
-	MqttServerURI      string      `json:"mqtt_server_uri"`
-	MqttUsername       string      `json:"mqtt_server_username"`
-	MqttPassword       string      `json:"mqtt_server_password"`
-	MqttClientIdPrefix string      `json:"mqtt_client_id_prefix"`
-	LogFile            string      `json:"log_file"`
-	LogLevel           string      `json:"log_level"`
-	LogFormat          string      `json:"log_format"`
-	ConfiguredAt       string      `json:"configured_at"`
-	ConfiguredBy       string      `json:"configured_by"`
-	CustomConfigs      interface{} `json:"custom_configs"`
+	WorkDir            string `json:"-"`
+	InstanceAddress    string `json:"instance_address"`
+	MqttServerURI      string `json:"mqtt_server_uri"`
+	MqttUsername       string `json:"mqtt_server_username"`
+	MqttPassword       string `json:"mqtt_server_password"`
+	MqttClientIdPrefix string `json:"mqtt_client_id_prefix"`
+	LogFile            string `json:"log_file"`
+	LogLevel           string `json:"log_level"`
+	LogFormat          string `json:"log_format"`
+	ConfiguredAt       string `json:"configured_at"`
+	ConfiguredBy       string `json:"configured_by"`
+	CustomConfigs      any    `json:"custom_configs"`
 }
 
 // NewConfigs stores main application configurations
@@ -51,7 +51,7 @@ func (cf *Configs) initFiles() error {
 }
 
 func (cf *Configs) LoadFromFile() error {
-	configFileBody, err := ioutil.ReadFile(cf.path)
+	configFileBody, err := os.ReadFile(cf.path)
 	if err != nil {
 		return err
 	}
@@ -66,7 +66,7 @@ func (cf *Configs) SaveToFile() error {
 	cf.ConfiguredBy = "auto"
 	cf.ConfiguredAt = time.Now().Format(time.RFC3339)
 	bpayload, err := json.Marshal(cf)
-	err = ioutil.WriteFile(cf.path, bpayload, 0664)
+	err = os.WriteFile(cf.path, bpayload, 0664)
 	if err != nil {
 		return err
 	}
@@ -84,18 +84,18 @@ func (cf *Configs) GetDefaultDir() string {
 func (cf *Configs) LoadDefaults() error {
 	configFile := filepath.Join(cf.WorkDir, "data", "config.json")
 	if err := os.Remove(configFile); err != nil {
-		log.Error(err)
+		log.Error("[edgeapp] ", err)
 	}
-	log.Info("Config file doesn't exist.Loading default config")
+	log.Info("[edgeapp] Config file doesn't exist.Loading default config")
 	defaultConfigFile := filepath.Join(cf.WorkDir, "defaults", "config.json")
 	return utils.CopyFile(defaultConfigFile, configFile)
 }
 
-func (cf *Configs) SetCustomConfigs(config interface{}) {
+func (cf *Configs) SetCustomConfigs(config any) {
 	cf.CustomConfigs = config
 }
 
-func (cf *Configs) GetCustomConfigs() interface{} {
+func (cf *Configs) GetCustomConfigs() any {
 	return cf.CustomConfigs
 }
 
