@@ -2,8 +2,11 @@ package primefimp
 
 import (
 	"errors"
-	"io/ioutil"
+	"fmt"
+	"os"
+	"runtime/debug"
 	"sync"
+	"sync/atomic"
 	"time"
 
 	log "github.com/sirupsen/logrus"
@@ -28,10 +31,9 @@ type ApiClient struct {
 	subFilters            map[string]NotifyFilter
 	inMsgChan             fimpgo.MessageCh
 	stopFlag              bool
-	isNotifyRouterStarted bool
+	isNotifyRouterStarted atomic.Bool
 	notifChMux            sync.RWMutex
 	isVincAppsSyncEnabled bool
-	isConnPoolEnabled     bool
 	cloudService          string
 	responsePayloadType   string
 	globalPrefix          string
@@ -60,7 +62,7 @@ func NewApiClient(clientID string, mqttTransport *fimpgo.MqttTransport, loadSite
 	}
 	if loadSiteIntoCache {
 		if err := api.ReloadSiteToCache(3); err != nil {
-			log.Error("<PF-API> Error reloading cache: ", err)
+			log.Error("[fimpgo] Error reloading cache: ", err)
 		}
 	}
 
@@ -94,9 +96,9 @@ func (mh *ApiClient) IsCacheEmpty() bool {
 // ValidateAndReloadSiteCache validates cache , if empty it makes one reload attempt. The method can be used for cache lazy loading.
 func (mh *ApiClient) ValidateAndReloadSiteCache() bool {
 	if mh.IsCacheEmpty() {
-		log.Debug("<PF-API> Empty site cache.Reloading...")
+		log.Debug("[fimpgo] Empty site cache.Reloading...")
 		if err := mh.ReloadSiteToCache(1); err != nil {
-			log.Error("<PF-API> Error reloading cache: ", err)
+			log.Error("[fimpgo] Error reloading cache: ", err)
 		}
 		if mh.IsCacheEmpty() {
 			return false
@@ -111,30 +113,30 @@ func (mh *ApiClient) ReloadSiteToCache(retry int) error {
 	var site *Site
 	var err error
 	for i := 1; i < retry; i++ {
-		log.Debug("<PF-API> Reloading site into the cache.Attempt ", i)
+		log.Debug("[fimpgo] Reloading site into the cache.Attempt ", i)
 		site, err = mh.GetSite(false)
 		if err == nil {
-			log.Debug("<PF-API> Site loaded successfully")
+			log.Debug("[fimpgo] Site loaded successfully")
 			break
 		}
-		log.Error("<PF-API> site sync error :", err.Error())
+		log.Error("[fimpgo] site sync error :", err.Error())
 		time.Sleep(time.Second * time.Duration(5*i))
 
 	}
 	if err != nil {
 		mh.isCacheEnabled = false
-		log.Errorf("<PF-API>: %s", err)
+		log.Errorf("[fimpgo]: %s", err)
 		return err
 	}
 	mh.isCacheEnabled = true
 	mh.siteCache = *site
-	log.Debug("<PF-API> Site info successfully loaded to cache")
+	log.Debug("[fimpgo] Site info successfully loaded to cache")
 	return nil
 }
 
 // LoadVincResponseFromFile Loads site from file . File should be in exactly the same format as vinculum response
 func (mh *ApiClient) LoadVincResponseFromFile(fileName string) error {
-	bSite, err := ioutil.ReadFile(fileName)
+	bSite, err := os.ReadFile(fileName)
 	if err != nil {
 		return err
 	}
@@ -177,36 +179,38 @@ func (mh *ApiClient) UnregisterChannel(channelId string) {
 
 func (mh *ApiClient) StartNotifyRouter() {
 	go func() {
-		mh.isNotifyRouterStarted = true
+		mh.isNotifyRouterStarted.Store(true)
 		for {
 			if mh.stopFlag {
 				break
 			}
 			mh.notifyRouter()
-			log.Info("<PF-API> Restarting notify router")
+			log.Info("[fimpgo] Restarting notify router")
 		}
-		log.Info("<PF-API> Notify router stopped ")
+
+		log.Info("[fimpgo] Notify router stopped")
 	}()
 }
 
 // Stop : This is destructor
 func (mh *ApiClient) Stop() {
 	mh.sClient.Stop()
-	if mh.isNotifyRouterStarted {
+	if mh.isNotifyRouterStarted.Load() {
 		mh.stopFlag = true
 		mh.inMsgChan <- &fimpgo.Message{}
 	}
 }
 
 // UpdateSite : Updates the internal cache according to notification message.
-func (mh *ApiClient) UpdateSite(notif *Notify) {
-	log.Tracef("Command: %s & Component:%s", notif.Cmd, notif.Component)
+func (mh *ApiClient) UpdateSite(notif *Notify) error {
+	log.Tracef("Command=%s & Component=%s", notif.Cmd, notif.Component)
 	if !mh.isVincAppsSyncEnabled {
 		if notif.Component != ComponentArea && notif.Component != ComponentDevice && notif.Component != ComponentThing && notif.Component != ComponentRoom {
-			log.Debugf("Component skipped")
-			return
+			log.Tracef("Component skipped")
+			return nil
 		}
 	}
+
 	switch notif.Cmd {
 	case CmdAdd:
 		switch notif.Component {
@@ -225,14 +229,16 @@ func (mh *ApiClient) UpdateSite(notif *Notify) {
 		case ComponentTimer:
 			mh.siteCache.AddTimer(notif.GetTimer())
 		default:
-			log.Errorf("Unknown Component:%s cannot be added.", notif.Component)
+			return fmt.Errorf("unknown Component=%s add", notif.Component)
 		}
 	case CmdDelete:
-		err := mh.siteCache.RemoveWithID(notif.Component, int(notif.Id.(float64)))
+		notifID, ok := notif.Id.(float64)
+		if !ok {
+			return fmt.Errorf("notify ID type assertion error")
+		}
+		err := mh.siteCache.RemoveWithID(notif.Component, int(notifID))
 		if err != nil {
-			log.Error(err)
-		} else {
-			log.Infof("%s with ID:%d is deleted", notif.Component, int(notif.Id.(float64)))
+			return fmt.Errorf("remove with ID err: %v", err)
 		}
 	case CmdEdit:
 		switch notif.Component {
@@ -249,66 +255,77 @@ func (mh *ApiClient) UpdateSite(notif *Notify) {
 		case ComponentShortcut:
 			mh.siteCache.UpdateShortcut(notif.GetShortcut())
 		default:
-			log.Error("Unknown component update occured. Report this as issue please")
+			return fmt.Errorf("unknown component=%s update", notif.Component)
 		}
 	case CmdSet:
 		switch notif.Component {
 		case ComponentRoom:
 		case ComponentHub:
+		default:
+			return fmt.Errorf("unknown component=%s set", notif.Component)
 		}
+	default:
+		return fmt.Errorf("unknown notify cmd=%s", notif.Cmd)
 	}
 
+	return nil
 }
 
 // Receives notify messages and forwards them using filtering
 func (mh *ApiClient) notifyRouter() {
 	defer func() {
 		if r := recover(); r != nil {
-			log.Error("<PF-API> notify router CRASHED with error :", r)
+			log.Errorf("[fimpgo] isChannelInterested crash %v", r)
+			log.Info(debug.Stack())
 		}
 	}()
 
 	mh.inMsgChan = make(fimpgo.MessageCh, 10)
 	mh.mqttTransport.RegisterChannel(mh.clientID, mh.inMsgChan)
 	if err := mh.mqttTransport.Subscribe(VincEventTopic); err != nil {
-		log.Error("<PF-API> error subscribing to the vinculum event topic: ", err)
+		log.Error("[fimpgo] Subscribe to vinculum event topic err:", err)
 	}
 
 	for msg := range mh.inMsgChan {
 		if mh.stopFlag {
 			break
 		}
+
 		if msg.Topic != VincEventTopic {
 			continue
 		}
+
 		notif, err := FimpToNotify(msg)
 		if err != nil {
-			log.Debug("<PF-API> Can't cast to Notify. Err:", err)
+			log.Warnf("[fimpgo] Cast %v to Notify err: %v", msg, err)
 			continue
-		} else {
-			mh.UpdateSite(notif)
-			if mh.isNotifyRouterStarted { // make sure notify router is started
-				mh.notifChMux.RLock()
-				for cid, nfCh := range mh.notifySubChannels { // check all subfilters
-					nfFilter, ok := mh.subFilters[cid]
-					var send bool
-					if ok {
-						if nfFilter.Cmd == notif.Cmd && nfFilter.Component == notif.Component {
-							send = true
-						}
-					} else {
+		}
+
+		if err := mh.UpdateSite(notif); err != nil {
+			log.Warnf("[fimpgo] Update site from notify err: %v", err)
+		}
+
+		if mh.isNotifyRouterStarted.Load() { // make sure notify router is started
+			mh.notifChMux.RLock()
+			for cid, nfCh := range mh.notifySubChannels { // check all subfilters
+				nfFilter, ok := mh.subFilters[cid]
+				var send bool
+				if ok {
+					if nfFilter.Cmd == notif.Cmd && nfFilter.Component == notif.Component {
 						send = true
 					}
-					if send {
-						select {
-						case nfCh <- *notif: // send notification to corresponding subchannel if there is match
-						default:
-							log.Warnf("<PF-API> Send channel %s is blocked ", cid)
-						}
+				} else {
+					send = true
+				}
+				if send {
+					select {
+					case nfCh <- *notif: // send notification to corresponding subchannel if there is match
+					default:
+						log.Warnf("[fimpgo] Send channel %s is blocked ", cid)
 					}
 				}
-				mh.notifChMux.RUnlock()
 			}
+			mh.notifChMux.RUnlock()
 		}
 	}
 }
@@ -337,8 +354,7 @@ func (mh *ApiClient) sendGetRequest(components []string) (*fimpgo.FimpMessage, e
 	return mh.sClient.SendReqRespFimp(reqAddr.Serialize(), responseAddress, msg, 5, true)
 }
 
-func (mh *ApiClient) sendSetRequest(component string, value interface{}) (*fimpgo.FimpMessage, error) {
-
+func (mh *ApiClient) sendSetRequest(component string, value any) (*fimpgo.FimpMessage, error) {
 	reqAddr := fimpgo.Address{MsgType: fimpgo.MsgTypeCmd, ResourceType: fimpgo.ResourceTypeApp, ResourceName: "vinculum", ResourceAddress: "1"}
 	respAddr := mh.responseAddress()
 	responseAddress := respAddr.Serialize()
