@@ -5,17 +5,16 @@ import (
 	"errors"
 
 	"github.com/futurehomeno/fimpgo"
-	log "github.com/sirupsen/logrus"
 )
 
 type Notify struct {
-	Errors     interface{}     `json:"errors"`
+	Errors     any             `json:"errors"`
 	Cmd        string          `json:"cmd"`
 	Component  string          `json:"component"`
 	ParamRaw   json.RawMessage `json:"param"`
 	ChangesRaw json.RawMessage `json:"changes"`
 	Success    bool            `json:"success"`
-	Id         interface{}     `json:"id,omitempty"`
+	Id         any             `json:"id,omitempty"`
 }
 
 type DeleteChange struct {
@@ -137,11 +136,9 @@ func (ntf *Notify) GetHub() *Hub {
 }
 
 func (ntf *Notify) GetDeleteChange() *DeleteChange {
-	log.Debug("Delete change requested")
 	var result DeleteChange
 	err := json.Unmarshal(ntf.ChangesRaw, &result)
 	if err != nil {
-		log.Error(err)
 		return nil
 	}
 	return &result
