@@ -2,18 +2,23 @@ package utils
 
 import (
 	"encoding/json"
-	"io/ioutil"
+	"os"
 )
 
 type TestConfig struct {
 	BrokerURI string
 	SiteId    string
-
 }
 
-func GetTestConfig(path string ) TestConfig {
+func GetTestConfig(path string) TestConfig {
 	var result TestConfig
-	bFile, _ := ioutil.ReadFile(path)
-	json.Unmarshal(bFile,&result)
+	bFile, err := os.ReadFile(path)
+	if err != nil {
+		panic(err)
+	}
+
+	if err = json.Unmarshal(bFile, &result); err != nil {
+		panic(err)
+	}
 	return result
 }
