@@ -12,7 +12,6 @@ const (
 	AppCurrentStateNotConfigured = "NOT_CONFIGURED"
 	AppCurrentStateRunning       = "RUNNING"
 	AppCurrentStateERROR         = "ERROR"
-
 )
 
 type Resource struct {
@@ -23,8 +22,8 @@ type Resource struct {
 	Author                 string            `json:"author"`
 	Version                string            `json:"version"`
 	PackageName            string            `json:"package_name"` // in some cases package may have different name from service/resource name
-	State                  string            `json:"state"`    // Current application state
-	AppInfo                AppInfo           `json:"app_info"` // Either App or Adapter , it's defined by ResourceType
+	State                  string            `json:"state"`        // Current application state
+	AppInfo                AppInfo           `json:"app_info"`     // Either App or Adapter , it's defined by ResourceType
 	AdapterInfo            AdapterInfo       `json:"adapter_info"`
 	ConfigRequired         bool              `json:"config_required"` // if true , the adapter should be configured before it can be used
 	Configs                map[string]string `json:"configs"`         // configurations params
@@ -33,16 +32,17 @@ type Resource struct {
 	IsInstanceConfigurable bool              `json:"is_instance_configurable"` // if true , the instance of adapter/app has to be configured before it can be used . false - adapter/app can be used without instance configuration
 	InstanceId             string            `json:"instance_id"`              // Some system configurations can allow to run multiple instances of the same app or adapter , for instance multiple hubs under the same site and with radio module every hub
 }
+
 // Will be removed in future
 type AppInfo struct {
 }
 
 // Will be removed in future
 type AdapterInfo struct {
-	FwVersion             string             `json:"fw_version"` // should be in Semantic Versioning format .
-	Technology            string             `json:"technology"`
-	HwDependency          map[string]string  `json:"hw_dependency"`           //  {"serialPort":"/dev/ttyUSB0"} ,
-	NetworkManagementType string             `json:"network_management_type"` // "inclusion_exclusion", "inclusion_dev_remove" , "full_sync"
+	FwVersion             string            `json:"fw_version"` // should be in Semantic Versioning format .
+	Technology            string            `json:"technology"`
+	HwDependency          map[string]string `json:"hw_dependency"`           //  {"serialPort":"/dev/ttyUSB0"} ,
+	NetworkManagementType string            `json:"network_management_type"` // "inclusion_exclusion", "inclusion_dev_remove" , "full_sync"
 }
 
 type ServiceDiscoveryResponder struct {
@@ -63,7 +63,11 @@ func NewServiceDiscoveryResponder(mqt *fimpgo.MqttTransport) *ServiceDiscoveryRe
 
 // Start responder service listener
 func (sr *ServiceDiscoveryResponder) Start() {
-	sr.mqt.Subscribe(sr.discoveryRequestTopic)
+	if err := sr.mqt.Subscribe(sr.discoveryRequestTopic); err != nil {
+		logrus.Error("[fimpgo]D iscovery responder subscribe err:", err)
+		return
+	}
+
 	sr.mqt.RegisterChannelWithFilter("discovery-responder", sr.requestsCh, struct {
 		Topic     string
 		Service   string
@@ -86,10 +90,11 @@ func (sr *ServiceDiscoveryResponder) responder() {
 	for {
 		select {
 		case <-sr.requestsCh:
-			logrus.Debug("New responder request")
 			msg := fimpgo.NewMessage("evt.discovery.report", "system", fimpgo.VTypeObject, sr.resource, nil, nil, nil)
 			adr := fimpgo.Address{MsgType: fimpgo.MsgTypeEvt, ResourceType: fimpgo.ResourceTypeDiscovery}
-			sr.mqt.Publish(&adr, msg)
+			if err := sr.mqt.Publish(&adr, msg); err != nil {
+				logrus.Error("[fimpgo] Discovery responder publish err:", err)
+			}
 		case <-sr.stopSignal:
 			break
 		}
