@@ -1,14 +1,18 @@
 package edgeapp
 
 import (
-log "github.com/sirupsen/logrus"
-"gopkg.in/natefinch/lumberjack.v2"
+	"github.com/futurehomeno/fimpgo/edgeapp/formatters"
+	log "github.com/sirupsen/logrus"
+	"gopkg.in/natefinch/lumberjack.v2"
 )
 
 func SetupLog(logfile string, level string, logFormat string) {
-	if logFormat == "json" {
+	switch logFormat {
+	case "json":
 		log.SetFormatter(&log.JSONFormatter{TimestampFormat: "2006-01-02 15:04:05.999"})
-	} else {
+	case "budzik":
+		log.SetFormatter(formatters.NewBudzikFormatter())
+	default:
 		log.SetFormatter(&log.TextFormatter{FullTimestamp: true, ForceColors: true, TimestampFormat: "2006-01-02T15:04:05.999"})
 	}
 
