@@ -1,210 +1,215 @@
 package fimpgo
 
 import (
-	MQTT "github.com/eclipse/paho.mqtt.golang"
-	log "github.com/sirupsen/logrus"
+	"math/rand"
+	"sync"
+	"sync/atomic"
 	"testing"
 	"time"
+
+	log "github.com/sirupsen/logrus"
+	"github.com/stretchr/testify/assert"
 )
-
-
 
 var msgChan = make(chan int)
 
-func onMsg(topic string, addr *Address, iotMsg *FimpMessage,rawMessage []byte){
-	log.Info("New message")
-	if addr.ServiceName == "temp_sensor" && addr.ServiceAddress == "300"{
+func onMsg(topic string, addr *Address, iotMsg *FimpMessage, rawMessage []byte) {
+	log.Infof("New msg %s val=%v", topic, iotMsg.Value)
+	if addr.ServiceName == "temp_sensor" && addr.ServiceAddress == "300" {
 		msgChan <- 1
-	}else {
+	} else {
 		msgChan <- 2
 	}
 }
 
 var isCorrect = make(map[int]bool)
 
-
-
 func TestMqttTransport_Publish(t *testing.T) {
-
 	log.SetLevel(log.DebugLevel)
-	mqtt := NewMqttTransport("tcp://localhost:1883","fimpgotest","","",true,1,1)
+	mqtt := NewMqttTransport("tcp://127.0.0.1:1883", "fimpgotest", "", "", true, 1, 1)
 	err := mqtt.Start()
-	t.Log("Connected")
 	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Fatal("Start MQTT err:", err)
 	}
 
 	mqtt.SetMessageHandler(onMsg)
-	mqtt.Subscribe("#")
-	t.Log("Publishing message")
+	if err := mqtt.Subscribe("#"); err != nil {
+		t.Fatal("Subscribe err:", err)
+	}
 
 	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", float64(35.5), nil, nil, nil)
 	adr := Address{MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
-	mqtt.Publish(&adr,msg)
+	mqtt.Publish(&adr, msg)
 
-	t.Log("Waiting for new message")
-	result := <- msgChan
-	t.Log("Got new message")
-	mqtt.Stop()
+	result := <-msgChan
+
 	if result != 1 {
-		t.Error("Wrong message")
+		t.Error("Wrong message result=", result)
 	}
 
+	mqtt.Stop()
 }
 
 func TestMqttTransport_PublishStopPublish(t *testing.T) {
-
 	log.SetLevel(log.DebugLevel)
-	mqtt := NewMqttTransport("tcp://localhost:1883","fimpgotest","","",true,1,1)
+	mqtt := NewMqttTransport("tcp://127.0.0.1:1883", "fimpgotest", "", "", true, 1, 1)
 	err := mqtt.Start()
-	t.Log("Connected")
 	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Fatal("Start MQTT err:", err)
 	}
 
 	mqtt.SetMessageHandler(onMsg)
-	mqtt.Subscribe("#")
-	t.Log("Publishing message")
+	if err := mqtt.Subscribe("#"); err != nil {
+		t.Fatal("Subscribe err:", err)
+	}
 
 	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", float64(35.5), nil, nil, nil)
 	adr := Address{MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
-	mqtt.Publish(&adr,msg)
+	mqtt.Publish(&adr, msg)
 
-	t.Log("Waiting for new message")
-	result := <- msgChan
-	t.Log("Got new message")
-	mqtt.Stop()
+	result := <-msgChan
 	if result != 1 {
-		t.Error("Wrong message")
-	}
-	time.Sleep(time.Second*5)
-	mqtt = NewMqttTransport("tcp://localhost:1883","fimpgotest","","",true,1,1)
-	err = mqtt.Start()
-	t.Log("Connected 2")
-	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Errorf("Wrong message result=%d", result)
 	}
 
-	time.Sleep(time.Second*5)
-
-	t.Log("Done")
 	mqtt.Stop()
 
-}
+	mqtt = NewMqttTransport("tcp://127.0.0.1:1883", "fimpgotest", "", "", true, 1, 1)
+	err = mqtt.Start()
+	if err != nil {
+		t.Fatal("Start MQTT err:", err)
+	}
 
+	mqtt.Stop()
+}
 
 func TestMqttTransport_PublishSync(t *testing.T) {
 	log.SetLevel(log.DebugLevel)
-	MQTT.DEBUG = log.StandardLogger()
-	mqtt := NewMqttTransport("tcp://localhost:1883","fimpgotest","","",true,1,1)
+	mqtt := NewMqttTransport("tcp://127.0.0.1:1883", "fimpgotest", "", "", true, 1, 1)
 	err := mqtt.Start()
-	t.Log("Connected")
 	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Fatal("Start MQTT err:", err)
 	}
 
-	t.Log("Publishing message")
+	var cnt atomic.Int64
 
-	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", float64(35.5), nil, nil, nil)
+	mqtt.SetMessageHandler(func(topic string, addr *Address, iotMsg *FimpMessage, rawMessage []byte) {
+		if addr.ServiceName != "temp_sensor" {
+			return
+		}
+
+		val, err := iotMsg.GetIntValue()
+		if err != nil {
+			log.Warnf("Wrong value %t err: %v", iotMsg.Value, err)
+		} else {
+			cnt.Add(val)
+		}
+	})
+
+	if err := mqtt.Subscribe("pt:j1/mt:evt/#"); err != nil {
+		t.Fatal("Subscribe err:", err)
+	}
+
+	msg := NewIntMessage("evt.sensor.report", "temp_sensor", int64(35), nil, nil, nil)
 	adr := Address{MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
 
-	for i:=0;i<10;i++ {
-		err = mqtt.PublishSync(&adr,msg)
-		if err != nil {
-			log.Info("Publish failed . Err :",)
-		}else {
-			log.Info("Publish success ")
-		}
-		time.Sleep(time.Second*5)
+	expVal := int64(0)
+	for range 10 {
+		msg.Value = int64(rand.Intn(100))
+		expVal += msg.Value.(int64)
 
+		err = mqtt.PublishSync(&adr, msg)
+		if err != nil {
+			log.Info("Publish failed err:", err)
+		}
 	}
 
-	t.Log("Waiting for new message")
-	t.Log("Got new message")
+	time.Sleep(200 * time.Millisecond)
 	mqtt.Stop()
 
+	assert.Equal(t, expVal, cnt.Load())
 }
 
 func TestMqttTransport_SubUnsub(t *testing.T) {
 	log.SetLevel(log.DebugLevel)
-	mqtt := NewMqttTransport("tcp://localhost:1883","fimpgotest","","",true,1,1)
+	mqtt := NewMqttTransport("tcp://127.0.0.1:1883", "fimpgotest", "", "", true, 1, 1)
 	err := mqtt.Start()
-	t.Log("Connected")
 	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Fatal("Start MQTT err:", err)
 	}
 
 	mqtt.SetMessageHandler(onMsg)
-	mqtt.Subscribe("pt:j1/mt:evt/#")
-	//mqtt.Subscribe("pt:j1/mt:evt/rt:dev/rn:test/ad:1/sv:temp_sensor/ad:300")
-	//mqtt.Unsubscribe("pt:j1/mt:evt/rt:dev/rn:test/ad:1/sv:temp_sensor/ad:300")
-	mqtt.Unsubscribe("pt:j1/mt:evt/#")
-	t.Log("Publishing message")
-
-	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", float64(35.5), nil, nil, nil)
-	adr := Address{PayloadType: DefaultPayload,MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
-	mqtt.PublishSync(&adr,msg)
-
-	t.Log("Waiting for new message")
-	result := <- msgChan
-	t.Log("Got new message")
-	mqtt.Stop()
-	if result != 1 {
-		t.Error("Wrong message")
+	if err := mqtt.Subscribe("pt:j1/mt:evt/#"); err != nil {
+		t.Fatal("Subscribe err:", err)
+		return
 	}
 
+	// unsubscribe and send message, shall not receive it
+	mqtt.Unsubscribe("pt:j1/mt:evt/#")
+
+	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", float64(35.5), nil, nil, nil)
+	adr := Address{PayloadType: DefaultPayload, MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
+	mqtt.PublishSync(&adr, msg)
+
+	select {
+	case <-msgChan:
+		t.Error("Should not receive msg")
+	case <-time.After(2 * time.Second):
+	}
+
+	mqtt.Stop()
 }
 
-func TestMqttTransport_PublishTls(t *testing.T) {
+// TODO: Fix, awsiot.private.key is not available in the repo
+/*func TestMqttTransport_PublishTls(t *testing.T) {
 	log.SetLevel(log.DebugLevel)
 	// for test replace XYZ with actual AWS IoT core address and ABC with actual clientid
-	mqtt := NewMqttTransport("ssl://a1ds8ixdqbiw53-ats.iot.eu-central-1.amazonaws.com:443","00000000alexdevtest","","",false,1,1)
+	mqtt := NewMqttTransport("ssl://a1ds8ixdqbiw53-ats.iot.eu-central-1.amazonaws.com:443", "00000000alexdevtest", "", "", false, 1, 1)
 
 	// for test enter valid site-id
 	mqtt.SetGlobalTopicPrefix("331D092F-4685-4CC9-8337-2598E6F5D8D5")
 	// for test place certificate and key into certs folder
-	err := mqtt.ConfigureTls("awsiot.private.key","awsiot.crt","./certs",true)
+	err := mqtt.ConfigureTls("awsiot.private.key", "awsiot.crt", "./certs", true)
 
 	if err != nil {
-		t.Error("Certificate error :",err)
+		t.Fatal("Configure TLS err", err)
 	}
 
 	err = mqtt.Start()
-	t.Log("Connected")
 	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Fatal("Start MQTT err:", err)
 	}
 
 	mqtt.SetMessageHandler(onMsg)
-	time.Sleep(time.Second*1)
-	mqtt.Subscribe("#")
-	t.Log("Publishing message")
+
+	if err := mqtt.Subscribe("#"); err != nil {
+		t.Fatal("Subscribe err:", err)
+	}
 
 	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", float64(35.5), nil, nil, nil)
-	adr := Address{PayloadType: DefaultPayload,MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
-	mqtt.Publish(&adr,msg)
+	adr := Address{PayloadType: DefaultPayload, MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
+	mqtt.Publish(&adr, msg)
 
-	t.Log("Waiting for new message")
-	result := <- msgChan
-	t.Log("Got new message")
-	mqtt.Stop()
+	result := <-msgChan
+
 	if result != 1 {
 		t.Error("Wrong message")
 	}
 
-}
+	mqtt.Stop()
+}*/
 
-func TestMqttTransport_PublishTls_2(t *testing.T) {
-
+// TODO: Fix, awsiot.private.key is not available in the repo
+/*func TestMqttTransport_PublishTls_2(t *testing.T) {
 	connConfig := MqttConnectionConfigs{
-		ServerURI:           "ssl://a1ds8ixdqbiw53-ats.iot.eu-central-1.amazonaws.com:443",
-		ClientID:            "00000000alexdevtest",
-		CleanSession:        true,
-		SubQos:              1,
-		PubQos:              1,
-		CertDir:             "./certs",
-		PrivateKeyFileName:  "awsiot.private.key",
-		CertFileName:        "awsiot.crt",
+		ServerURI:          "ssl://a1ds8ixdqbiw53-ats.iot.eu-central-1.amazonaws.com:443",
+		ClientID:           "00000000alexdevtest",
+		CleanSession:       true,
+		SubQos:             1,
+		PubQos:             1,
+		CertDir:            "./certs",
+		PrivateKeyFileName: "awsiot.private.key",
+		CertFileName:       "awsiot.crt",
 	}
 
 	log.SetLevel(log.DebugLevel)
@@ -214,276 +219,309 @@ func TestMqttTransport_PublishTls_2(t *testing.T) {
 	// for test enter valid site-id
 	mqtt.SetGlobalTopicPrefix("331D092F-4685-4CC9-8337-2598E6F5D8D5")
 	// for test place certificate and key into certs folder
-	err := mqtt.ConfigureTls("awsiot.private.key","awsiot.crt","./certs",true)
+	err := mqtt.ConfigureTls("awsiot.private.key", "awsiot.crt", "./certs", true)
 
 	if err != nil {
-		t.Error("Certificate error :",err)
+		t.Fatal("Configure TLS err:", err)
 	}
 
 	err = mqtt.Start()
-	t.Log("Connected")
 	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Fatal("Start MQTT err:", err)
 	}
 
 	mqtt.SetMessageHandler(onMsg)
-	time.Sleep(time.Second*1)
-	mqtt.Subscribe("#")
-	t.Log("Publishing message")
+
+	if err := mqtt.Subscribe("#"); err != nil {
+		t.Fatal("Subscribe err:", err)
+	}
 
 	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", float64(35.5), nil, nil, nil)
-	adr := Address{PayloadType: DefaultPayload,MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
-	mqtt.Publish(&adr,msg)
+	adr := Address{PayloadType: DefaultPayload, MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
+	mqtt.Publish(&adr, msg)
 
-	t.Log("Waiting for new message")
-	result := <- msgChan
-	t.Log("Got new message")
-	mqtt.Stop()
+	result := <-msgChan
 	if result != 1 {
 		t.Error("Wrong message")
 	}
 
-}
-
+	mqtt.Stop()
+}*/
 
 func TestMqttTransport_TestChannels(t *testing.T) {
-
 	log.SetLevel(log.DebugLevel)
-	mqtt := NewMqttTransport("tcp://localhost:1883","fimpgotest","","",true,1,1)
+	mqtt := NewMqttTransport("tcp://127.0.0.1:1883", "fimpgotest", "", "", true, 1, 1)
 	err := mqtt.Start()
-	t.Log("Connected")
-	time.Sleep(time.Second*1)
-	mqtt.Subscribe("#")
 	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Fatal("Start MQTT err:", err)
 	}
+
+	if err := mqtt.Subscribe("#"); err != nil {
+		t.Fatal("Subscribe err:", err)
+	}
+
 	chan1 := make(MessageCh)
 	chan2 := make(MessageCh)
-	mqtt.RegisterChannel("chan1",chan1)
-	mqtt.RegisterChannel("chan2",chan2)
-	isCorrect[1] = false
-	isCorrect[2] = false
+	mqtt.RegisterChannel("chan1", chan1)
+	mqtt.RegisterChannel("chan2", chan2)
+	correctMsg := make(chan int, 2)
+
+	var wg sync.WaitGroup
+	wg.Add(2)
+
 	go func(msgChan MessageCh) {
-	   newMsg :=<- chan1
-	   if newMsg.Payload.Service == "temp_sensor" {
-	   		isCorrect[1] = true
-	   }
-	}(chan1)
-	go func(msgChan MessageCh) {
-		newMsg :=<- chan2
+		wg.Done()
+		newMsg := <-chan1
 		if newMsg.Payload.Service == "temp_sensor" {
-			isCorrect[2] = true
+			correctMsg <- 1
+		}
+	}(chan1)
+
+	go func(msgChan MessageCh) {
+		wg.Done()
+		newMsg := <-chan2
+		if newMsg.Payload.Service == "temp_sensor" {
+			correctMsg <- 2
 		}
 	}(chan2)
 
+	wg.Wait()
+
 	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", float64(35.5), nil, nil, nil)
-	adr := Address{PayloadType: DefaultPayload,MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
-	mqtt.Publish(&adr,msg)
-	time.Sleep(time.Second*1)
+	adr := Address{PayloadType: DefaultPayload, MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
+	mqtt.Publish(&adr, msg)
+
+	expVals := map[int]bool{1: true, 2: true}
+
+	for len(expVals) > 0 {
+		select {
+		case val := <-correctMsg:
+			delete(expVals, val)
+
+		case <-time.After(2 * time.Second):
+			t.Fatal("Message not received within timeout missing:", expVals)
+		}
+	}
+
 	mqtt.UnregisterChannel("chan1")
 	mqtt.UnregisterChannel("chan2")
-	if isCorrect[1] && isCorrect[2] {
-		t.Log("Channel test - OK")
-	}else {
-		t.Error("Wrong result")
-		t.Fail()
-	}
+	mqtt.Stop()
 }
 
 func TestMqttTransport_TestResponder(t *testing.T) {
-
 	log.SetLevel(log.TraceLevel)
-	var isResponseReceived bool
-	mqtt := NewMqttTransport("tcp://localhost:1883","fimpgotest-1","","",true,1,1)
+
+	mqtt := NewMqttTransport("tcp://127.0.0.1:1883", "fimpgotest-1", "", "", true, 1, 1)
 	err := mqtt.Start()
-	t.Log("Connected")
-	time.Sleep(time.Second*1)
-	mqtt.Subscribe("#")
+	if err != nil {
+		t.Fatal("Start MQTT err:", err)
+	}
 
-	mqtt2 := NewMqttTransport("tcp://localhost:1883","fimpgotest-2","","",true,1,1)
+	if err := mqtt.Subscribe("#"); err != nil {
+		t.Fatal("Subscribe err:", err)
+	}
+
+	mqtt2 := NewMqttTransport("tcp://127.0.0.1:1883", "fimpgotest-2", "", "", true, 1, 1)
 	err = mqtt2.Start()
-	t.Log("Connected")
-	time.Sleep(time.Second*1)
-	mqtt2.Subscribe("pt:j1c1/mt:rsp/rt:app/rn:response_tester/ad:1")
-
 
 	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Fatal("Start MQTT 2 err:", err)
 	}
+
+	if err := mqtt2.Subscribe("pt:j1c1/mt:rsp/rt:app/rn:response_tester/ad:1"); err != nil {
+		t.Fatal("Subscribe response_tester err:", err)
+	}
+
 	chan1 := make(MessageCh)
 	chan2 := make(MessageCh)
-	mqtt.RegisterChannel("chan1",chan1)
-	mqtt2.RegisterChannel("chan2",chan2)
-	// responder
-	go func(msgChan MessageCh) {
-		for {
-			newMsg :=<- chan1
-			t.Log("New message for responder. Service = "+newMsg.Payload.Service)
-			if newMsg.Payload.Service == "tester" {
-				mqtt.RespondToRequest(newMsg.Payload,NewFloatMessage("evt.test.response", "test_responder", 35.5, nil, nil, nil))
-			}
-		}
+	mqtt.RegisterChannel("chan1", chan1)
+	mqtt2.RegisterChannel("chan2", chan2)
 
+	var wg sync.WaitGroup
+	wg.Add(2)
+
+	go func(msgChan MessageCh) {
+		wg.Done()
+		newMsg := <-chan1
+		if newMsg.Payload.Service == "tester" {
+			if err := mqtt.RespondToRequest(newMsg.Payload, NewFloatMessage("evt.test.response", "test_responder", 35.5, nil, nil, nil)); err != nil {
+				t.Error("Error responding to request:", err)
+			}
+			return
+		}
 	}(chan1)
 
+	rspReceived := make(chan struct{})
+
 	go func(msgChan MessageCh) {
-		for {
-			newMsg :=<- chan2
-			t.Log("Service = "+newMsg.Payload.Service)
-			if newMsg.Payload.Service == "test_responder" && newMsg.Topic == "pt:j1c1/mt:rsp/rt:app/rn:response_tester/ad:1" {
-				isResponseReceived = true
-			}
+		wg.Done()
+		newMsg := <-chan2
+		if newMsg.Payload.Service == "test_responder" && newMsg.Topic == "pt:j1c1/mt:rsp/rt:app/rn:response_tester/ad:1" {
+			close(rspReceived)
+			return
+		} else {
+			t.Error("Wrong response message received :", newMsg)
 		}
-
-
 	}(chan2)
+
+	wg.Wait()
 
 	msg := NewFloatMessage("cmd.test.get_response", "tester", float64(35.5), nil, nil, nil)
 	msg.ResponseToTopic = "pt:j1c1/mt:rsp/rt:app/rn:response_tester/ad:1"
-	adr := Address{PayloadType: DefaultPayload,MsgType: MsgTypeCmd, ResourceType: ResourceTypeApp, ResourceName: "test", ResourceAddress: "1"}
-	mqtt.Publish(&adr,msg)
-	time.Sleep(time.Second*2)
+	adr := Address{PayloadType: DefaultPayload, MsgType: MsgTypeCmd, ResourceType: ResourceTypeApp, ResourceName: "test", ResourceAddress: "1"}
+	mqtt.Publish(&adr, msg)
+
+	select {
+	case <-rspReceived:
+
+	case <-time.After(3 * time.Second):
+		t.Error("Response not received within timeout")
+		t.Fail()
+	}
+
 	mqtt.UnregisterChannel("chan1")
 	mqtt.UnregisterChannel("chan2")
 	mqtt.Unsubscribe("#")
-	if isResponseReceived {
-		t.Log("Response received")
-	}else {
-		t.Error("Wrong result")
-		t.Fail()
-	}
+	mqtt.Stop()
 }
 
 func TestMqttTransport_TestChannelsWithFilters(t *testing.T) {
-
 	log.SetLevel(log.DebugLevel)
-	mqtt := NewMqttTransport("tcp://localhost:1883","fimpgotest","","",true,1,1)
+	mqtt := NewMqttTransport("tcp://127.0.0.1:1883", "fimpgotest", "", "", true, 1, 1)
 	err := mqtt.Start()
-	t.Log("Connected")
-	time.Sleep(time.Second*1)
-	mqtt.Subscribe("#")
 	if err != nil {
-		t.Error("Error connecting to broker ",err)
+		t.Fatal("Start MQTT err:", err)
+		return
 	}
+
+	if err := mqtt.Subscribe("#"); err != nil {
+		t.Fatal("Subscribe err:", err)
+		return
+	}
+
 	chan1 := make(MessageCh)
 	chan2 := make(MessageCh)
 	chan3 := make(MessageCh)
 	chan4 := make(MessageCh)
 	chan5 := make(MessageCh)
-	mqtt.RegisterChannel("chan1",chan1)
-	mqtt.RegisterChannel("chan2",chan2)
-	mqtt.RegisterChannelWithFilter("chan3",chan3,FimpFilter{
+	mqtt.RegisterChannel("chan1", chan1)
+	mqtt.RegisterChannel("chan2", chan2)
+	mqtt.RegisterChannelWithFilter("chan3", chan3, FimpFilter{
 		Topic:     "pt:j1/mt:evt/#",
 		Service:   "temp_sensor",
 		Interface: "evt.sensor.report",
 	})
 
-	mqtt.RegisterChannelWithFilter("chan4",chan4,FimpFilter{
+	mqtt.RegisterChannelWithFilter("chan4", chan4, FimpFilter{
 		Topic:     "pt:j1/mt:cmd/#",
 		Service:   "temp_sensor",
 		Interface: "cmd.sensor.report",
 	})
 
-	testFilterFunc := func (topic string, addr *Address, iotMsg *FimpMessage) bool {
-		if iotMsg.Type == "evt.sensor.report"{
-			return true
-		}
-		return false
+	testFilterFunc := func(topic string, addr *Address, iotMsg *FimpMessage) bool {
+		return iotMsg.Type == "evt.sensor.report"
 	}
 
-	mqtt.RegisterChannelWithFilterFunc("chan5",chan5,testFilterFunc)
+	mqtt.RegisterChannelWithFilterFunc("chan5", chan5, testFilterFunc)
 
-	isCorrect[1] = false
-	isCorrect[2] = false
-	isCorrect[3] = false
-	isCorrect[4] = true
-	isCorrect[5] = false
+	var startedWg sync.WaitGroup
+	startedWg.Add(5)
+	correctMsg := make(chan int, 2)
+
 	go func(msgChan MessageCh) {
-		newMsg :=<- msgChan
+		startedWg.Done()
+		newMsg := <-msgChan
 		if newMsg.Payload.Service == "temp_sensor" {
-			isCorrect[1] = true
+			correctMsg <- 1
 		}
 	}(chan1)
+
 	go func(msgChan MessageCh) {
-		newMsg :=<- msgChan
+		startedWg.Done()
+		newMsg := <-msgChan
 		if newMsg.Payload.Service == "temp_sensor" {
-			isCorrect[2] = true
+			correctMsg <- 2
 		}
 	}(chan2)
 
 	go func(msgChan MessageCh) {
-		newMsg :=<- msgChan
+		startedWg.Done()
+		newMsg := <-msgChan
 		if newMsg.Payload.Service == "temp_sensor" {
-			isCorrect[3] = true
+			correctMsg <- 3
 		}
 	}(chan3)
-	// Negative test 
-	go func(msgChan MessageCh) {
-		_=<- msgChan
-		isCorrect[4] = false
 
+	go func(msgChan MessageCh) {
+		startedWg.Done()
+		<-msgChan
+		correctMsg <- 4
 	}(chan4)
 
 	go func(msgChan MessageCh) {
-		_=<- msgChan
-		isCorrect[5] = true
-
+		startedWg.Done()
+		<-msgChan
+		correctMsg <- 5
 	}(chan5)
+
+	startedWg.Wait()
 
 	msg := NewFloatMessage("evt.sensor.report", "temp_sensor", float64(35.5), nil, nil, nil)
 	adr := Address{MsgType: MsgTypeEvt, ResourceType: ResourceTypeDevice, ResourceName: "test", ResourceAddress: "1", ServiceName: "temp_sensor", ServiceAddress: "300"}
-	mqtt.Publish(&adr,msg)
-	time.Sleep(time.Second*1)
+	mqtt.Publish(&adr, msg)
+
+	expVals := map[int]bool{1: true, 2: true, 3: true, 5: true}
+
+	for len(expVals) > 0 {
+		select {
+		case val := <-correctMsg:
+			if val == 4 {
+				t.Error("Should not receive msg on chan4")
+				t.Fail()
+				return
+			}
+
+			delete(expVals, val)
+
+		case <-time.After(2 * time.Second):
+			t.Fatal("Message not received within timeout missing:", expVals)
+		}
+	}
+
 	mqtt.UnregisterChannel("chan1")
 	mqtt.UnregisterChannel("chan2")
 	mqtt.UnregisterChannel("chan3")
 	mqtt.UnregisterChannel("chan4")
 	mqtt.UnregisterChannel("chan5")
-	if isCorrect[1] && isCorrect[2] && isCorrect[3] && isCorrect[4] && isCorrect[5]{
-		t.Log("Channel test - OK")
-	}else {
-		t.Error("Wrong result")
-		t.Log(isCorrect)
-		t.Fail()
-	}
-
+	mqtt.Stop()
 
 }
 
 func TestAddGlobalPrefixToTopic(t *testing.T) {
-	result := AddGlobalPrefixToTopic("12345","pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
+	result := AddGlobalPrefixToTopic("12345", "pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
 	if result != "12345/pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0" {
 		t.Error("Wrong topic")
-	}else {
-		t.Log("AddGlobalPrefixToTopic test 1 - OK")
 	}
-	result = AddGlobalPrefixToTopic("12345","/pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
+
+	result = AddGlobalPrefixToTopic("12345", "/pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
 	if result != "12345/pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0" {
 		t.Error("Wrong topic")
-	}else {
-		t.Log("AddGlobalPrefixToTopic test 2 - OK")
 	}
-	result = AddGlobalPrefixToTopic("","pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
+
+	result = AddGlobalPrefixToTopic("", "pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
 	if result != "pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0" {
 		t.Error("Wrong topic")
-	}else {
-		t.Log("AddGlobalPrefixToTopic test 3 - OK")
 	}
 }
 
 func TestDetachGlobalPrefixFromTopic(t *testing.T) {
-	globalPrefix,topic := DetachGlobalPrefixFromTopic("12345/pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
+	globalPrefix, topic := DetachGlobalPrefixFromTopic("12345/pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
 	if globalPrefix != "12345" || topic != "pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0" {
 		t.Error("Wrong topic")
-	}else {
-		t.Log("DetachGlobalPrefixFromTopic test 1 - OK")
 	}
-	globalPrefix,topic = DetachGlobalPrefixFromTopic("ABC/12345/pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
+
+	globalPrefix, topic = DetachGlobalPrefixFromTopic("ABC/12345/pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0")
 	if globalPrefix != "ABC/12345" || topic != "pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:dev_sys/ad:49_0" {
 		t.Error("Wrong topic")
-	}else {
-		t.Log("Result ,",globalPrefix,topic)
-		t.Log("DetachGlobalPrefixFromTopic test 2 - OK")
 	}
 }
-
