@@ -1,6 +1,7 @@
 package fimpgo
 
 import (
+	"fmt"
 	"time"
 
 	"github.com/google/uuid"
@@ -9,16 +10,13 @@ import (
 
 // SyncClient allows sync interaction over async channel.
 type SyncClient struct {
-	mqttTransport         *MqttTransport
-	mqttConnPool          *MqttConnectionPool
-	isConnPoolEnabled     bool
-	transactionPoolSize   int // Max transaction pool size
-	inboundBufferSize     int // Inbound message channel buffer size
-	inboundMsgChannel     MessageCh
-	inboundChannelName    string
-	stopSignalCh          chan bool
-	isStartedUsingConnect bool
-	globalPrefix          string
+	mqttTransport       *MqttTransport
+	mqttConnPool        *MqttConnectionPool
+	isConnPoolEnabled   bool
+	transactionPoolSize int // Max transaction pool size
+	inboundBufferSize   int // Inbound message channel buffer size
+	stopSignalCh        chan bool
+	globalPrefix        string
 }
 
 // SetGlobalPrefix configures global prefix/site_id . Most be used from backend services.
@@ -52,9 +50,13 @@ func (sc *SyncClient) SetConfigs(transactionPoolSize int, inboundBuffSize int) {
 	if transactionPoolSize == 0 {
 		transactionPoolSize = 20
 	}
+
 	if inboundBuffSize == 0 {
 		inboundBuffSize = 10
 	}
+
+	sc.transactionPoolSize = transactionPoolSize
+	sc.inboundBufferSize = inboundBuffSize
 }
 
 func (sc *SyncClient) init() {
@@ -65,16 +67,11 @@ func (sc *SyncClient) init() {
 // Should be used if MqttTransport instance is not provided in constructor .
 func (sc *SyncClient) Connect(serverURI string, clientID string, username string, password string, cleanSession bool, subQos byte, pubQos byte) error {
 	if sc.mqttTransport == nil {
-		log.Info("<SyncClient> Connecting to mqtt broker")
 		sc.mqttTransport = NewMqttTransport(serverURI, clientID, username, password, cleanSession, subQos, pubQos)
 		err := sc.mqttTransport.Start()
 		if err != nil {
-			log.Error("<SyncClient> Error connecting to broker :", err)
 			return err
 		}
-		sc.isStartedUsingConnect = true
-	} else {
-		log.Info("<SyncClient> Already connected")
 	}
 
 	return nil
@@ -82,40 +79,31 @@ func (sc *SyncClient) Connect(serverURI string, clientID string, username string
 
 // Stop has to be invoked to stop message listener
 func (sc *SyncClient) Stop() {
-	if sc.isStartedUsingConnect {
-		sc.mqttTransport.Stop()
-	}
-
+	sc.mqttTransport.Stop()
 }
 
 // AddSubscription has to be invoked before Send methods
-func (sc *SyncClient) AddSubscription(topic string) {
-	if err := sc.mqttTransport.Subscribe(topic); err != nil {
-		log.Error("<SyncClient> error subscribing to topic:", err)
-	}
+func (sc *SyncClient) AddSubscription(topic string) error {
+	return sc.mqttTransport.Subscribe(topic)
 }
 
 // RemoveSubscription
-func (sc *SyncClient) RemoveSubscription(topic string) {
-	if err := sc.mqttTransport.Unsubscribe(topic); err != nil {
-		log.Error("<SyncClient> error unsubscribing from topic:", err)
-	}
+func (sc *SyncClient) RemoveSubscription(topic string) error {
+	return sc.mqttTransport.Unsubscribe(topic)
 }
 
 // SendFimpWithTopicResponse send message over mqtt and awaits response from responseTopic with responseService and responseMsgType
 func (sc *SyncClient) sendFimpWithTopicResponse(topic string, fimpMsg *FimpMessage, responseTopic string, responseService string, responseMsgType string, timeout int64, autoSubscribe bool) (*FimpMessage, error) {
-	//log.Debug("Registering request uid = ",fimpMsg.UID)
 	var conId int
 	var conn *MqttTransport
 	var inboundCh = make(MessageCh, 10)
-	var responseChannel chan *FimpMessage
 	var err error
 	var chanName = uuid.New().String()
 
 	defer func() {
 		if autoSubscribe && responseTopic != "" && conn != nil {
 			if err := conn.Unsubscribe(responseTopic); err != nil {
-				log.Error("<SyncClient> error unsubscribing from topic:", err)
+				log.Error("[fimpgo] Error unsubscribing from topic:", err)
 			}
 		}
 		if conn != nil {
@@ -137,9 +125,10 @@ func (sc *SyncClient) sendFimpWithTopicResponse(topic string, fimpMsg *FimpMessa
 	} else {
 		conn = sc.mqttTransport
 	}
+
 	conn.RegisterChannel(chanName, inboundCh)
 
-	responseChannel = sc.startResponseListener(fimpMsg, responseMsgType, responseService, responseTopic, inboundCh, timeout)
+	responseChannel := sc.startResponseListener(fimpMsg, responseMsgType, responseService, responseTopic, inboundCh, timeout)
 
 	// force the global prefix -> this is useful for per-site operations
 	if sc.globalPrefix != "" {
@@ -147,26 +136,23 @@ func (sc *SyncClient) sendFimpWithTopicResponse(topic string, fimpMsg *FimpMessa
 	}
 
 	if autoSubscribe && responseTopic != "" {
-		if err := conn.Subscribe(responseTopic); err != nil {
-			log.Error("<SyncClient> error subscribing to topic:", err)
+		if err = conn.Subscribe(responseTopic); err != nil {
+			return nil, fmt.Errorf("[fimpgo] Subscribe err: %s", err)
 		}
 	} else if responseTopic != "" {
-		if err := conn.Subscribe(responseTopic); err != nil {
-			log.Error("<SyncClient> error subscribing to topic:", err)
-			return nil, errSubscribe
+		if err = conn.Subscribe(responseTopic); err != nil {
+			return nil, fmt.Errorf("[fimpgo] Subscribe err: %s", err)
 		}
 	}
 
-	if err := conn.PublishToTopic(topic, fimpMsg); err != nil {
-		log.Error("<SyncClient> error publishing to topic:", err)
-		return nil, errPublish
+	if err = conn.PublishToTopic(topic, fimpMsg); err != nil {
+		return nil, fmt.Errorf("[fimpgo] Publish err: %s", err)
 	}
 
 	select {
 	case fimpResponse := <-responseChannel:
 		return fimpResponse, nil
 	case <-time.After(time.Second * time.Duration(timeout)):
-		log.Info("<SyncClient> No response from queue for ", timeout)
 		return nil, errTimeout
 	}
 }
@@ -189,8 +175,8 @@ func (sc *SyncClient) SendFimpWithTopicResponse(topic string, fimpMsg *FimpMessa
 
 // startResponseListener starts response listener , it blocks callers proc until response is received or timeout.
 func (sc *SyncClient) startResponseListener(requestMsg *FimpMessage, respMsgType, respService, respTopic string, inboundCh MessageCh, timeout int64) chan *FimpMessage {
-	log.Debug("<SyncClient> Msg listener is started")
 	respChan := make(chan *FimpMessage)
+
 	go func() {
 		for msg := range inboundCh {
 			if (respMsgType == msg.Payload.Type && respService == msg.Payload.Service && respTopic == msg.Topic) || requestMsg.UID == msg.Payload.CorrelationID {
