@@ -2,14 +2,15 @@ package utils
 
 import (
 	"errors"
-	log "github.com/sirupsen/logrus"
 	"net"
 )
 
 var (
-	errNoAvailableInterface = errors.New("no available interface")
-	errNoAvailableAddress   = errors.New("no available address")
+	errNoAvailableInterface  = errors.New("no available interface")
+	errNoAvailableAddress    = errors.New("no available address")
+	errNotSupportedInterface = errors.New("not supported interface")
 )
+
 // RoutedInterface returns a network interface that can route IP
 // traffic and satisfies flags.
 //
@@ -18,7 +19,7 @@ func RoutedInterface(network string, flags net.Flags) (*net.Interface, error) {
 	switch network {
 	case "ip", "ip4", "ip6":
 	default:
-		return nil, errNoAvailableInterface
+		return nil, errNotSupportedInterface
 	}
 	ift, err := net.Interfaces()
 	if err != nil {
@@ -33,9 +34,8 @@ func RoutedInterface(network string, flags net.Flags) (*net.Interface, error) {
 		}
 		return &ifi, nil
 	}
-	return nil, errNoAvailableInterface
+	return nil, errNoAvailableAddress
 }
-
 
 func hasRoutableIP(network string, ifi *net.Interface) (net.IP, bool) {
 	ifat, err := ifi.Addrs()
@@ -43,7 +43,6 @@ func hasRoutableIP(network string, ifi *net.Interface) (net.IP, bool) {
 		return nil, false
 	}
 	for _, ifa := range ifat {
-		log.Debug("Intf name ",ifa.String())
 		switch ifa := ifa.(type) {
 		case *net.IPAddr:
 			if ip, ok := routableIP(network, ifa.IP); ok {
@@ -62,7 +61,7 @@ func routableIP(network string, ip net.IP) (net.IP, bool) {
 	if !ip.IsLoopback() && !ip.IsLinkLocalUnicast() && !ip.IsGlobalUnicast() {
 		return nil, false
 	}
-	log.Debug("IP is routable ",ip.IsLoopback())
+
 	switch network {
 	case "ip4":
 		if ip := ip.To4(); ip != nil {
